refact: extrair métodos de ataque padrão para skill-default-attacks.mjs (Fase 3 — Commit 1/4)
Remove ~1031 linhas de skill-manager.mjs extraindo os 5 métodos privados de ataque/suporte
para module/skills/skill-default-attacks.mjs como funções exportadas de módulo:
- performUnifiedSkillAttack (único ponto de entrada via #handleDefaultButtonClick)
- performDefaultPrimaryAttack / performDefaultSecondaryAttack (preservados, não chamados)
- checkAndConsumeAmmunition / detectCriticalResults (helpers puros)

skill-manager.mjs: 2287 → 1256 linhas. Sem mudança de comportamento.
