+++ module/skills/skill-manager.mjs
@@ -1,6 +1,7 @@
 import { BaseSkill } from './base-skill.mjs';
 import { buildRollFormula } from '../helpers/config.mjs';
 import { ChatMessageHelper } from '../helpers/chat-messages.mjs';
+import { checkAndConsumeAmmunition, detectCriticalResults, performDefaultPrimaryAttack, performDefaultSecondaryAttack, performUnifiedSkillAttack } from './skill-default-attacks.mjs';
 
 /**
  * Skill Manager - Orchestrates all skill-related functionality
@@ -537,7 +538,7 @@ export class SkillManager {
     switch (buttonType) {
       case 'attack':
         // Unified attack - hand selection via dialog checkboxes
-        await this.#performUnifiedSkillAttack(actor, skillName);
+        await performUnifiedSkillAttack(actor, skillName);
         break;
         
       case 'expand':
@@ -879,1037 +880,6 @@ export class SkillManager {
     } catch (error) {
       console.error("Erro ao abrir dialog de aplicação de efeitos:", error);
       ui.notifications.error("Erro ao abrir dialog de efeitos. Verifique o console para mais detalhes.");
-    }
-  }
-
-  /**
-   * Perform a default primary attack for unregistered skills
-   * @param {Actor} actor - The actor
-   * @param {string} skillName - The skill name
-   * @private
-   */
-  static async #performDefaultPrimaryAttack(actor, skillName) {
-    try {
-      // Validate target selection
-      const selectedTargets = game.user.targets;
-      if (selectedTargets.size === 0) {
-        ui.notifications.warn("Por favor, selecione um ou mais alvos para atacar.");
-        return;
-      }
-
-      // Get all REAL weapons (not virtual unarmed attacks) that are equipped
-      const realWeapons = actor.items.filter(item => 
-        item.type === 'arma' && 
-        item.system.equipped && 
-        !item.system.isUnarmed &&
-        (item.system.rightHand || item.system.leftHand)
-      );
-
-      // Check if right hand is occupied by a REAL weapon
-      const primaryWeapon = realWeapons.find(weapon => weapon.system.rightHand);
-      
-      // Calculate weapon damage (base damage + ability modifiers)
-      let weaponDamage = 0;
-      if (primaryWeapon) {
-        const baseDamage = parseInt(primaryWeapon.system.damage?.value) || 0;
-        weaponDamage = baseDamage;
-        
-        // Add ability modifiers
-        if (primaryWeapon.system.damage?.useStrength) {
-          weaponDamage += actor.system.abilities.strength.value || 0;
-        }
-        if (primaryWeapon.system.damage?.useDexterity) {
-          weaponDamage += actor.system.abilities.dexterity.value || 0;
-        }
-      } else {
-        // No weapon equipped - calculate unarmed attack damage
-        const strengthValue = actor.system.abilities.strength.value || 0;
-        const strengthBonus = actor.system.abilities.strength.totalBonus || 0;
-        const totalStrength = strengthValue + strengthBonus;
-        weaponDamage = totalStrength > 0 ? totalStrength : 1; // Minimum 1 damage
-      }
-      
-      // Check ammunition for ranged weapons before showing dialog
-      if (primaryWeapon && primaryWeapon.system.ranged) {
-        const canAttack = await this.#checkAndConsumeAmmunition(actor, primaryWeapon);
-        if (!canAttack) return;
-      }
-
-      // Import the advantage selection dialog
-      const { AdvantageSelectionDialog } = await import('../applications/advantage-selection-dialog.mjs');
-      
-      // Show advantage selection dialog
-      const result = await AdvantageSelectionDialog.show();
-      if (!result) return; // User cancelled
-
-      const { rollType, attackMode, manualModifier = 0, primaryHand, secondaryHand } = result;
-
-      // JOINT ROLL: Require selected tokens and hand selection
-      if (attackMode === 'conjunto') {
-        if (!canvas.tokens.controlled || canvas.tokens.controlled.length === 0) {
-          ui.notifications.warn('Por favor, selecione um ou mais tokens antes de fazer uma Rolagem em Conjunto.');
-          return;
-        }
-        if (!primaryHand && !secondaryHand) {
-          ui.notifications.warn('Por favor, selecione a Mão Primária ou a Mão Secundária para fazer uma Rolagem em Conjunto.');
-          return;
-        }
-      }
-
-      // Get roll data
-      const rollData = actor.getRollData();
-      
-      const formula = buildRollFormula(rollType, "@accuracy.total", manualModifier);
-      let rollDescription = "";
-      
-      switch (rollType) {
-        case 'advantage':
-          rollDescription = "Rolagem com Vantagem";
-          break;
-        case 'disadvantage':
-          rollDescription = "Rolagem com Desvantagem";
-          break;
-        case 'enhanced-advantage':
-          rollDescription = "Rolagem com Vantagem Aprimorada";
-          break;
-        case 'enhanced-disadvantage':
-          rollDescription = "Rolagem com Desvantagem Aprimorada";
-          break;
-        case 'normal':
-          rollDescription = "Rolagem Normal";
-          break;
-        default:
-          return;
-      }
-
-      // Check if we need to make individual attacks for each target
-      const shouldRollIndividually = attackMode === 'individual' && selectedTargets.size > 1;
-      
-      if (shouldRollIndividually) {
-        // Make individual attack for each target
-        const targetArray = Array.from(selectedTargets);
-        
-        // Show notification about multiple attacks
-        ui.notifications.info(`Realizando ${targetArray.length} ataques individuais...`);
-        
-        for (let i = 0; i < targetArray.length; i++) {
-          const targetToken = targetArray[i];
-          
-          // Add small delay between attacks for visual clarity
-          if (i > 0) {
-            await new Promise(resolve => setTimeout(resolve, 300));
-          }
-          
-          // Add attack mode and target name to description
-          let individualRollDescription = `${rollDescription} (Individual) → ${targetToken.name}`;
-          
-          // Check for Congelado effect and apply skill penalty
-          const { CongeladoEffect } = await import('../effects/effects/congelado.mjs');
-          const congeladoPenalty = CongeladoEffect.getSkillPenalty(actor);
-          
-          // Apply Congelado penalty to formula if present
-          if (congeladoPenalty !== 0) {
-            formula += ` ${congeladoPenalty}`;
-            individualRollDescription += ` [Congelado ${congeladoPenalty}]`;
-          }
-          
-          // Create flavor text
-          const flavorText = `<div style="text-align: center; margin-bottom: 4px;">
-            <strong>${skillName}</strong> - ${individualRollDescription}
-          </div>`;
-          
-          // Roll with critical detection
-          const roll = new Roll(formula, rollData);
-          await roll.evaluate();
-          
-          // Apply Sangramento effect for accuracy rolls
-          const { SangramentoEffect } = await import('../effects/effects/sangramento.mjs');
-          await SangramentoEffect.applyBleedingDamage(actor, 'Precisão', 'accuracy');
-          
-          // Detect critical results using accuracy logic
-          const flags = this.#detectCriticalResults(roll, actor, 'accuracy');
-          
-          // Calculate final damage (double on critical hit)
-          const isCriticalHit = flags?.cardigan?.criticalHit || false;
-          const finalDamage = isCriticalHit ? weaponDamage * 2 : weaponDamage;
-          
-          // Show notification for critical results (only for the user who rolled)
-          if (flags?.cardigan?.criticalHit) {
-            const critThreshold = actor.system?.details?.criticalHit;
-            if (critThreshold) {
-              ui.notifications.info(`Acerto Crítico! (${roll.total} >= ${critThreshold})`);
-            } else {
-              ui.notifications.info(`Acerto Crítico!`);
-            }
-          } else if (flags?.cardigan?.criticalFailure) {
-            // Check if weapon will lose durability
-            if (primaryWeapon && primaryWeapon.system.durability) {
-              const currentDurability = primaryWeapon.system.durability.current;
-              if (currentDurability > 0) {
-                const newDurability = Math.max(0, currentDurability - 1);
-                ui.notifications.warn(`Erro Crítico! ${primaryWeapon.name} perdeu durabilidade (${currentDurability} → ${newDurability})`);
-              } else {
-                ui.notifications.warn(`Erro Crítico!`);
-              }
-            } else {
-              ui.notifications.warn(`Erro Crítico!`);
-            }
-          }
-          
-          // Handle critical failure - reduce weapon durability
-          const isCriticalFailure = flags.cardigan?.criticalFailure || false;
-          if (isCriticalFailure && primaryWeapon && primaryWeapon.system.durability) {
-            const currentDurability = primaryWeapon.system.durability.current;
-            if (currentDurability > 0) {
-              const newDurability = Math.max(0, currentDurability - 1);
-              await primaryWeapon.update({
-                'system.durability.current': newDurability
-              });
-            }
-          }
-          
-          // Add single target data to flags
-          if (targetToken.actor) {
-            flags.cardigan = flags.cardigan || {};
-            flags.cardigan.attackTargets = {
-              targets: [{
-                tokenId: targetToken.id,
-                actorId: targetToken.actor.id,
-                name: targetToken.name
-              }],
-              attackerId: actor.id,
-              attackerName: actor.name,
-              skillName: skillName,
-              weaponId: primaryWeapon?._id || primaryWeapon?.id,
-              weaponName: primaryWeapon?.name,
-              weaponProperties: primaryWeapon?.system?.properties || [],
-              damage: weaponDamage,  // ALWAYS use BASE damage (not doubled) in flags
-              attackerCriticalHit: isCriticalHit  // Add critical hit flag
-            };
-          }
-          
-          // Determine roll mode: if current user is GM, use blind mode
-          const rollMode = game.settings.get('core', 'rollMode');
-
-          // Create message data with flags
-          const messageData = {
-            speaker: { alias: actor.name },
-            flavor: flavorText,
-            rolls: [roll],
-            flags: flags
-          };
-
-          // Apply roll mode using Foundry's official API method
-          ChatMessage.applyRollMode(messageData, rollMode);
-          
-          // Send attack roll to chat
-          await ChatMessage.create(messageData);
-        }
-        
-        return; // Exit after processing all individual attacks
-      }
-      
-      // Single attack for all targets (conjunto mode or single target)
-      const modeText = attackMode === 'conjunto' ? ' (Conjunto)' : ' (Individual)';
-      rollDescription += modeText;
-
-      // Check for Congelado effect and apply skill penalty
-      const { CongeladoEffect } = await import('../effects/effects/congelado.mjs');
-      const congeladoPenalty = CongeladoEffect.getSkillPenalty(actor);
-      
-      // Apply Congelado penalty to formula if present
-      if (congeladoPenalty !== 0) {
-        formula += ` ${congeladoPenalty}`;
-        rollDescription += ` [Congelado ${congeladoPenalty}]`;
-      }
-
-      // Create flavor text
-      const flavorText = `<div style="text-align: center; margin-bottom: 4px;">
-        <strong>${skillName}</strong> - ${rollDescription}
-      </div>`;
-
-      // Roll with critical detection
-      const roll = new Roll(formula, rollData);
-      await roll.evaluate();
-      
-      // Apply Sangramento effect for accuracy rolls
-      const { SangramentoEffect } = await import('../effects/effects/sangramento.mjs');
-      await SangramentoEffect.applyBleedingDamage(actor, 'Precisão', 'accuracy');
-
-      // Detect critical results using accuracy logic
-      const flags = this.#detectCriticalResults(roll, actor, 'accuracy');
-
-      // Calculate final damage (double on critical hit)
-      const isCriticalHit = flags?.cardigan?.criticalHit || false;
-      const finalDamage = isCriticalHit ? weaponDamage * 2 : weaponDamage;
-
-      // Show notification for critical results (only for the user who rolled)
-      if (flags?.cardigan?.criticalHit) {
-        const critThreshold = actor.system?.details?.criticalHit;
-        if (critThreshold) {
-          ui.notifications.info(`Acerto Crítico! (${roll.total} >= ${critThreshold})`);
-        } else {
-          ui.notifications.info(`Acerto Crítico!`);
-        }
-      } else if (flags?.cardigan?.criticalFailure) {
-        // Check if weapon will lose durability
-        if (primaryWeapon && primaryWeapon.system.durability) {
-          const currentDurability = primaryWeapon.system.durability.current;
-          if (currentDurability > 0) {
-            const newDurability = Math.max(0, currentDurability - 1);
-            ui.notifications.warn(`Erro Crítico! ${primaryWeapon.name} perdeu durabilidade (${currentDurability} → ${newDurability})`);
-          } else {
-            ui.notifications.warn(`Erro Crítico!`);
-          }
-        } else {
-          ui.notifications.warn(`Erro Crítico!`);
-        }
-      }
-
-      // Handle critical failure - reduce weapon durability
-      const isCriticalFailure = flags.cardigan?.criticalFailure || false;
-      if (isCriticalFailure && primaryWeapon && primaryWeapon.system.durability) {
-        const currentDurability = primaryWeapon.system.durability.current;
-        if (currentDurability > 0) {
-          const newDurability = Math.max(0, currentDurability - 1);
-          await primaryWeapon.update({
-            'system.durability.current': newDurability
-          });
-        }
-      }
-
-      // Collect target data for evasion buttons
-      const attackTargets = game.user.targets;
-      const targetData = [];
-      attackTargets.forEach(target => {
-        if (target.actor) {
-          targetData.push({
-            tokenId: target.id,
-            actorId: target.actor.id,
-            name: target.name
-          });
-        }
-      });
-
-      // Add target data to flags
-      if (targetData.length > 0) {
-        flags.cardigan = flags.cardigan || {};
-        flags.cardigan.attackTargets = {
-          targets: targetData,
-          attackerId: actor.id,
-          attackerName: actor.name,
-          skillName: skillName,
-          weaponId: primaryWeapon?._id || primaryWeapon?.id,
-          weaponName: primaryWeapon?.name,
-          weaponProperties: primaryWeapon?.system?.properties || [],
-          damage: weaponDamage,  // ALWAYS use BASE damage (not doubled) in flags
-          attackerCriticalHit: isCriticalHit  // Add critical hit flag
-        };
-      }
-
-      // Use player's roll mode setting (GM can choose blind manually)
-      const rollMode = game.settings.get('core', 'rollMode');
-
-      // Create message data with flags
-      const messageData = {
-        speaker: { alias: actor.name },
-        flavor: flavorText,
-        rolls: [roll],
-        flags: flags
-      };
-
-      // Apply roll mode using Foundry's official API method
-      ChatMessage.applyRollMode(messageData, rollMode);
-      
-      // Send attack roll to chat
-      await ChatMessage.create(messageData);
-
-    } catch (error) {
-      console.error(`Error performing default primary attack for ${skillName}:`, error);
-      ui.notifications.error(`Erro ao realizar ataque: ${error.message}`);
-    }
-  }
-
-  /**
-   * Perform a default secondary attack for unregistered skills
-   * @param {Actor} actor - The actor
-   * @param {string} skillName - The skill name
-   * @private
-   */
-  static async #performDefaultSecondaryAttack(actor, skillName) {
-    try {
-      // Validate target selection
-      const targets = game.user.targets;
-      if (targets.size === 0) {
-        ui.notifications.warn("Por favor, selecione um ou mais alvos para atacar.");
-        return;
-      }
-
-      // Get all REAL weapons (not virtual unarmed attacks) that are equipped
-      const realWeapons = actor.items.filter(item => 
-        item.type === 'arma' && 
-        item.system.equipped && 
-        !item.system.isUnarmed &&
-        (item.system.rightHand || item.system.leftHand)
-      );
-
-      // Find secondary hand weapon (leftHand only, NOT ambidextrous)
-      const secondaryWeapon = realWeapons.find(weapon => 
-        weapon.system.leftHand && !weapon.system.rightHand
-      );
-      
-      // Calculate weapon damage (base damage + ability modifiers)
-      let weaponDamage = 0;
-      if (secondaryWeapon) {
-        const baseDamage = parseInt(secondaryWeapon.system.damage?.value) || 0;
-        weaponDamage = baseDamage;
-        
-        // Add ability modifiers
-        if (secondaryWeapon.system.damage?.useStrength) {
-          weaponDamage += actor.system.abilities.strength.value || 0;
-        }
-        if (secondaryWeapon.system.damage?.useDexterity) {
-          weaponDamage += actor.system.abilities.dexterity.value || 0;
-        }
-      } else {
-        // No weapon equipped - calculate unarmed attack damage
-        const strengthValue = actor.system.abilities.strength.value || 0;
-        const strengthBonus = actor.system.abilities.strength.totalBonus || 0;
-        const totalStrength = strengthValue + strengthBonus;
-        weaponDamage = totalStrength > 0 ? totalStrength : 1; // Minimum 1 damage
-      }
-      
-      // Check ammunition for ranged weapons before showing dialog
-      if (secondaryWeapon && secondaryWeapon.system.ranged) {
-        const canAttack = await this.#checkAndConsumeAmmunition(actor, secondaryWeapon);
-        if (!canAttack) return;
-      }
-
-      // Import the advantage selection dialog
-      const { AdvantageSelectionDialog } = await import('../applications/advantage-selection-dialog.mjs');
-      
-      // Show advantage selection dialog
-      const result = await AdvantageSelectionDialog.show();
-      if (!result) return; // User cancelled
-
-      const { rollType, attackMode, manualModifier = 0, primaryHand, secondaryHand } = result;
-
-      // JOINT ROLL: Require multiple targets and hand selection
-      if (attackMode === 'conjunto') {
-        if (!game.user.targets || game.user.targets.size < 2) {
-          ui.notifications.warn('Por favor, selecione dois ou mais alvos antes de fazer uma Rolagem em Conjunto.');
-          return;
-        }
-        if (!primaryHand && !secondaryHand) {
-          ui.notifications.warn('Por favor, selecione a Mão Primária ou a Mão Secundária para fazer uma Rolagem em Conjunto.');
-          return;
-        }
-      }
-
-      // Get roll data
-      const rollData = actor.getRollData();
-      
-      const formula = buildRollFormula(rollType, "@accuracy.total", manualModifier);
-      let rollDescription = "";
-      
-      switch (rollType) {
-        case 'advantage':
-          rollDescription = "Rolagem com Vantagem";
-          break;
-        case 'disadvantage':
-          rollDescription = "Rolagem com Desvantagem";
-          break;
-        case 'enhanced-advantage':
-          rollDescription = "Rolagem com Vantagem Aprimorada";
-          break;
-        case 'enhanced-disadvantage':
-          rollDescription = "Rolagem com Desvantagem Aprimorada";
-          break;
-        case 'normal':
-          rollDescription = "Rolagem Normal";
-          break;
-        default:
-          return;
-      }
-
-      // Check if we need to make individual attacks for each target
-      const shouldRollIndividually = attackMode === 'individual' && targets.size > 1;
-      
-      if (shouldRollIndividually) {
-        // Make individual attack for each target
-        const targetArray = Array.from(targets);
-        
-        // Show notification about multiple attacks
-        ui.notifications.info(`Realizando ${targetArray.length} ataques secundários individuais...`);
-        
-        for (let i = 0; i < targetArray.length; i++) {
-          const targetToken = targetArray[i];
-          
-          // Add small delay between attacks for visual clarity
-          if (i > 0) {
-            await new Promise(resolve => setTimeout(resolve, 300));
-          }
-          
-          // Add attack mode and target name to description
-          const individualRollDescription = `${rollDescription} (Individual) → ${targetToken.name}`;
-          
-          // Create flavor text
-          const flavorText = `<div style="text-align: center; margin-bottom: 4px;">
-            <strong>${skillName}</strong> (Secundário) - ${individualRollDescription}
-          </div>`;
-          
-          // Roll with critical detection
-          const roll = new Roll(formula, rollData);
-          await roll.evaluate();
-          
-          // Apply Sangramento effect for accuracy rolls
-          const { SangramentoEffect } = await import('../effects/effects/sangramento.mjs');
-          await SangramentoEffect.applyBleedingDamage(actor, 'Precisão', 'accuracy');
-          
-          // Detect critical results using accuracy logic
-          const flags = this.#detectCriticalResults(roll, actor, 'accuracy');
-          
-          // Calculate final damage (double on critical hit)
-          const isCriticalHit = flags?.cardigan?.criticalHit || false;
-          const finalDamage = isCriticalHit ? weaponDamage * 2 : weaponDamage;
-          
-          // Show notification for critical results (only for the user who rolled)
-          if (flags?.cardigan?.criticalHit) {
-            const critThreshold = actor.system?.details?.criticalHit;
-            if (critThreshold) {
-              ui.notifications.info(`Acerto Crítico! (${roll.total} >= ${critThreshold})`);
-            } else {
-              ui.notifications.info(`Acerto Crítico!`);
-            }
-          } else if (flags?.cardigan?.criticalFailure) {
-            // Check if weapon will lose durability
-            if (secondaryWeapon && secondaryWeapon.system.durability) {
-              const currentDurability = secondaryWeapon.system.durability.current;
-              if (currentDurability > 0) {
-                const newDurability = Math.max(0, currentDurability - 1);
-                ui.notifications.warn(`Erro Crítico! ${secondaryWeapon.name} perdeu durabilidade (${currentDurability} → ${newDurability})`);
-              } else {
-                ui.notifications.warn(`Erro Crítico!`);
-              }
-            } else {
-              ui.notifications.warn(`Erro Crítico!`);
-            }
-          }
-          
-          // Handle critical failure - reduce weapon durability
-          const isCriticalFailure = flags.cardigan?.criticalFailure || false;
-          if (isCriticalFailure && secondaryWeapon && secondaryWeapon.system.durability) {
-            const currentDurability = secondaryWeapon.system.durability.current;
-            if (currentDurability > 0) {
-              const newDurability = Math.max(0, currentDurability - 1);
-              await secondaryWeapon.update({
-                'system.durability.current': newDurability
-              });
-            }
-          }
-          
-          // Add single target data to flags
-          if (targetToken.actor) {
-            flags.cardigan = flags.cardigan || {};
-            flags.cardigan.attackTargets = {
-              targets: [{
-                tokenId: targetToken.id,
-                actorId: targetToken.actor.id,
-                name: targetToken.name
-              }],
-              attackerId: actor.id,
-              attackerName: actor.name,
-              skillName: skillName,
-              weaponId: secondaryWeapon?._id || secondaryWeapon?.id,
-              weaponName: secondaryWeapon?.name,
-              weaponProperties: secondaryWeapon?.system?.properties || [],
-              damage: weaponDamage,  // ALWAYS use BASE damage (not doubled) in flags
-              attackerCriticalHit: isCriticalHit  // Add critical hit flag
-            };
-          }
-          
-          // Determine roll mode: if current user is GM, use blind mode
-          const rollMode = game.settings.get('core', 'rollMode');
-
-          // Create message data with flags
-          const messageData = {
-            speaker: { alias: actor.name },
-            flavor: flavorText,
-            rolls: [roll],
-            flags: flags
-          };
-
-          // Apply roll mode using Foundry's official API method
-          ChatMessage.applyRollMode(messageData, rollMode);
-          
-          // Send attack roll to chat
-          await ChatMessage.create(messageData);
-        }
-        
-        return; // Exit after processing all individual attacks
-      }
-      
-      // Single attack for all targets (conjunto mode or single target)
-      const modeText = attackMode === 'conjunto' ? ' (Conjunto)' : ' (Individual)';
-      rollDescription += modeText;
-
-      // Create flavor text
-      const flavorText = `<div style="text-align: center; margin-bottom: 4px;">
-        <strong>${skillName}</strong> (Secundário) - ${rollDescription}
-      </div>`;
-
-      // Roll with critical detection
-      const roll = new Roll(formula, rollData);
-      await roll.evaluate();
-      
-      // Apply Sangramento effect for accuracy rolls
-      const { SangramentoEffect } = await import('../effects/effects/sangramento.mjs');
-      await SangramentoEffect.applyBleedingDamage(actor, 'Precisão', 'accuracy');
-
-      // Detect critical results using accuracy logic
-      const flags = this.#detectCriticalResults(roll, actor, 'accuracy');
-
-      // Calculate final damage (double on critical hit)
-      const isCriticalHit = flags?.cardigan?.criticalHit || false;
-      const finalDamage = isCriticalHit ? weaponDamage * 2 : weaponDamage;
-
-      // Show notification for critical results (only for the user who rolled)
-      if (flags?.cardigan?.criticalHit) {
-        const critThreshold = actor.system?.details?.criticalHit;
-        if (critThreshold) {
-          ui.notifications.info(`Acerto Crítico! (${roll.total} >= ${critThreshold})`);
-        } else {
-          ui.notifications.info(`Acerto Crítico!`);
-        }
-      } else if (flags?.cardigan?.criticalFailure) {
-        // Check if weapon will lose durability
-        if (secondaryWeapon && secondaryWeapon.system.durability) {
-          const currentDurability = secondaryWeapon.system.durability.current;
-          if (currentDurability > 0) {
-            const newDurability = Math.max(0, currentDurability - 1);
-            ui.notifications.warn(`Erro Crítico! ${secondaryWeapon.name} perdeu durabilidade (${currentDurability} → ${newDurability})`);
-          } else {
-            ui.notifications.warn(`Erro Crítico!`);
-          }
-        } else {
-          ui.notifications.warn(`Erro Crítico!`);
-        }
-      }
-
-      // Handle critical failure - reduce weapon durability
-      const isCriticalFailure = flags.cardigan?.criticalFailure || false;
-      if (isCriticalFailure && secondaryWeapon && secondaryWeapon.system.durability) {
-        const currentDurability = secondaryWeapon.system.durability.current;
-        if (currentDurability > 0) {
-          const newDurability = Math.max(0, currentDurability - 1);
-          await secondaryWeapon.update({
-            'system.durability.current': newDurability
-          });
-        }
-      }
-
-      // Collect target data for evasion buttons
-      const attackTargets = game.user.targets;
-      const targetData = [];
-      attackTargets.forEach(target => {
-        if (target.actor) {
-          targetData.push({
-            tokenId: target.id,
-            actorId: target.actor.id,
-            name: target.name
-          });
-        }
-      });
-
-      // Add target data to flags
-      if (targetData.length > 0) {
-        flags.cardigan = flags.cardigan || {};
-        flags.cardigan.attackTargets = {
-          targets: targetData,
-          attackerId: actor.id,
-          attackerName: actor.name,
-          skillName: skillName,
-          weaponId: secondaryWeapon?._id || secondaryWeapon?.id,
-          weaponName: secondaryWeapon?.name,
-          weaponProperties: secondaryWeapon?.system?.properties || [],
-          damage: weaponDamage,  // ALWAYS use BASE damage (not doubled) in flags
-          attackerCriticalHit: isCriticalHit  // Add critical hit flag
-        };
-      }
-
-      // Use player's roll mode setting (GM can choose blind manually)
-      const rollMode = game.settings.get('core', 'rollMode');
-
-      // Create message data with flags
-      const messageData = {
-        speaker: { alias: actor.name },
-        flavor: flavorText,
-        rolls: [roll],
-        flags: flags
-      };
-
-      // Apply roll mode using Foundry's official API method
-      ChatMessage.applyRollMode(messageData, rollMode);
-      
-      // Send attack roll to chat
-      await ChatMessage.create(messageData);
-
-    } catch (error) {
-      console.error(`Error performing default secondary attack for ${skillName}:`, error);
-      ui.notifications.error(`Erro ao realizar ataque: ${error.message}`);
-    }
-  }
-
-  /**
-   * Perform a unified attack for skills - weapon selection based on hand checkboxes
-   * @param {Actor} actor - The actor
-   * @param {string} skillName - The skill name
-   * @private
-   */
-  static async #performUnifiedSkillAttack(actor, skillName) {
-    try {
-      // Validate target selection
-      const selectedTargets = game.user.targets;
-      if (selectedTargets.size === 0) {
-        ui.notifications.warn("Por favor, selecione um ou mais alvos para atacar.");
-        return;
-      }
-
-      // Import the advantage selection dialog
-      const { AdvantageSelectionDialog } = await import('../applications/advantage-selection-dialog.mjs');
-      
-      // Show advantage selection dialog with hand checkboxes visible
-      const result = await AdvantageSelectionDialog.show();
-      if (!result) return; // User cancelled
-
-      const { rollType, attackMode, manualModifier = 0, primaryHand, secondaryHand } = result;
-
-      // JOINT ROLL: Require multiple targets and hand selection
-      if (attackMode === 'conjunto') {
-        if (!game.user.targets || game.user.targets.size < 2) {
-          ui.notifications.warn('Por favor, selecione dois ou mais alvos antes de fazer uma Rolagem em Conjunto.');
-          return;
-        }
-        if (!primaryHand && !secondaryHand) {
-          ui.notifications.warn('Por favor, selecione a Mão Primária ou a Mão Secundária para fazer uma Rolagem em Conjunto.');
-          return;
-        }
-      }
-
-      // Determine which weapon to use based on checkbox selection
-      let selectedWeapon = null;
-      let weaponSource = null; // 'primary', 'secondary', or 'unarmed'
-      
-      // Get all REAL weapons (not virtual unarmed attacks) that are equipped
-      const realWeapons = actor.items.filter(item => 
-        item.type === 'arma' && 
-        item.system.equipped && 
-        !item.system.isUnarmed &&
-        (item.system.rightHand || item.system.leftHand)
-      );
-
-      // Priority: Primary hand first, then secondary hand
-      if (primaryHand) {
-        // Use right hand weapon
-        selectedWeapon = realWeapons.find(weapon => weapon.system.rightHand);
-        weaponSource = 'primary';
-      } else if (secondaryHand) {
-        // Use left hand weapon
-        selectedWeapon = realWeapons.find(weapon => weapon.system.leftHand);
-        weaponSource = 'secondary';
-      } else {
-        // No checkbox selected - default to primary hand or unarmed
-        selectedWeapon = realWeapons.find(weapon => weapon.system.rightHand);
-        weaponSource = selectedWeapon ? 'primary' : 'unarmed';
-      }
-      
-      // Calculate weapon damage (base damage + ability modifiers)
-      let weaponDamage = 0;
-      if (selectedWeapon) {
-        const baseDamage = parseInt(selectedWeapon.system.damage?.value) || 0;
-        weaponDamage = baseDamage;
-        
-        // Add ability modifiers
-        if (selectedWeapon.system.damage?.useStrength) {
-          weaponDamage += actor.system.abilities.strength.value || 0;
-        }
-        if (selectedWeapon.system.damage?.useDexterity) {
-          weaponDamage += actor.system.abilities.dexterity.value || 0;
-        }
-      } else {
-        // No weapon equipped - calculate unarmed attack damage
-        const strengthValue = actor.system.abilities.strength.value || 0;
-        const strengthBonus = actor.system.abilities.strength.totalBonus || 0;
-        const totalStrength = strengthValue + strengthBonus;
-        weaponDamage = totalStrength > 0 ? totalStrength : 1; // Minimum 1 damage
-        weaponSource = 'unarmed';
-      }
-      
-      // Check ammunition for ranged weapons
-      if (selectedWeapon && selectedWeapon.system.ranged) {
-        const canAttack = await this.#checkAndConsumeAmmunition(actor, selectedWeapon);
-        if (!canAttack) return;
-      }
-
-      // Get roll data
-      const rollData = actor.getRollData();
-      
-      const formula = buildRollFormula(rollType, "@accuracy.total", manualModifier);
-      let rollDescription = "";
-      
-      switch (rollType) {
-        case 'advantage':
-          rollDescription = "Rolagem com Vantagem";
-          break;
-        case 'disadvantage':
-          rollDescription = "Rolagem com Desvantagem";
-          break;
-        case 'enhanced-advantage':
-          rollDescription = "Rolagem com Vantagem Aprimorada";
-          break;
-        case 'enhanced-disadvantage':
-          rollDescription = "Rolagem com Desvantagem Aprimorada";
-          break;
-        case 'normal':
-          rollDescription = "Rolagem Normal";
-          break;
-        default:
-          return;
-      }
-
-      // Add attack mode to description
-      const modeText = attackMode === 'conjunto' ? ' (Conjunto)' : ' (Individual)';
-      rollDescription += modeText;
-      
-      // Add weapon source indicator
-      let weaponIndicator = '';
-      if (weaponSource === 'primary') {
-        weaponIndicator = ' [Mão Primária]';
-      } else if (weaponSource === 'secondary') {
-        weaponIndicator = ' [Mão Secundária]';
-      } else if (weaponSource === 'unarmed') {
-        weaponIndicator = ' [Desarmado]';
-      }
-      rollDescription += weaponIndicator;
-
-      // Check for Congelado effect and apply skill penalty
-      const { CongeladoEffect } = await import('../effects/effects/congelado.mjs');
-      const congeladoPenalty = CongeladoEffect.getSkillPenalty(actor);
-      
-      // Collect modifiers for display
-      const modifiers = [];
-      
-      // Apply Congelado penalty to formula if present
-      if (congeladoPenalty !== 0) {
-        formula += ` ${congeladoPenalty}`;
-        modifiers.push(`Congelado ${congeladoPenalty}`);
-      }
-      
-      // Roll with critical detection
-      const roll = new Roll(formula, rollData);
-      await roll.evaluate();
-      
-      // Apply Sangramento effect for accuracy rolls
-      const { SangramentoEffect } = await import('../effects/effects/sangramento.mjs');
-      await SangramentoEffect.applyBleedingDamage(actor, 'Precisão', 'accuracy');
-      
-      // Detect critical results using accuracy logic
-      const flags = this.#detectCriticalResults(roll, actor, 'accuracy');
-      
-      // Calculate final damage (double on critical hit)
-      const isCriticalHit = flags?.cardigan?.criticalHit || false;
-      const isCriticalFailure = flags?.cardigan?.criticalFailure || false;
-      const finalDamage = isCriticalHit ? weaponDamage * 2 : weaponDamage;
-      
-      // Show notification for critical results
-      if (isCriticalHit) {
-        const critThreshold = actor.system?.details?.criticalHit;
-        if (critThreshold) {
-          ui.notifications.info(`Acerto Crítico! (${roll.total} >= ${critThreshold})`);
-        } else {
-          ui.notifications.info(`Acerto Crítico!`);
-        }
-      } else if (isCriticalFailure) {
-        if (selectedWeapon && selectedWeapon.system.durability) {
-          const currentDurability = selectedWeapon.system.durability.current;
-          if (currentDurability > 0) {
-            const newDurability = Math.max(0, currentDurability - 1);
-            ui.notifications.warn(`Erro Crítico! ${selectedWeapon.name} perdeu durabilidade (${currentDurability} → ${newDurability})`);
-          } else {
-            ui.notifications.warn(`Erro Crítico!`);
-          }
-        } else {
-          ui.notifications.warn(`Erro Crítico!`);
-        }
-      }
-      
-      // Handle critical failure - reduce weapon durability
-      if (isCriticalFailure && selectedWeapon && selectedWeapon.system.durability) {
-        const currentDurability = selectedWeapon.system.durability.current;
-        if (currentDurability > 0) {
-          const newDurability = Math.max(0, currentDurability - 1);
-          await selectedWeapon.update({
-            'system.durability.current': newDurability
-          });
-        }
-      }
-      
-      // Collect target data for evasion buttons
-      const targetData = [];
-      selectedTargets.forEach(target => {
-        if (target.actor) {
-          targetData.push({
-            tokenId: target.id,
-            actorId: target.actor.id,
-            name: target.name
-          });
-        }
-      });
-
-      // Add target data to flags
-      if (targetData.length > 0) {
-        flags.cardigan = flags.cardigan || {};
-        flags.cardigan.attackTargets = {
-          targets: targetData,
-          attackerId: actor.id,
-          attackerName: actor.name,
-          skillName: skillName,
-          weaponId: selectedWeapon?._id || selectedWeapon?.id,
-          weaponName: selectedWeapon?.name,
-          weaponProperties: selectedWeapon?.system?.properties || [],
-          damage: weaponDamage,  // ALWAYS use BASE damage (not doubled) in flags
-          attackerCriticalHit: isCriticalHit  // Add critical hit flag
-        };
-      }
-
-      // Use player's roll mode setting
-      const rollMode = game.settings.get('core', 'rollMode');
-
-      // Send to chat using helper
-      await ChatMessageHelper.createRollMessage({
-        actor: actor,
-        roll: roll,
-        label: 'PRECISÃO',
-        rollType: rollType,
-        rollDescription: rollDescription,
-        handIndicator: null,
-        modifiers: modifiers,
-        flags: flags,
-        rollMode: rollMode,
-        isJointRoll: attackMode === 'conjunto',
-        primaryHand: primaryHand,
-        secondaryHand: secondaryHand
-      });
-
-    } catch (error) {
-      console.error(`Error performing unified skill attack for ${skillName}:`, error);
-      ui.notifications.error(`Erro ao realizar ataque: ${error.message}`);
-    }
-  }
-
-  /**
-   * Check ammunition and consume it for ranged weapons
-   * @param {Actor} actor - The actor
-   * @param {Item} weapon - The weapon item
-   * @returns {Promise<boolean>} True if attack can proceed, false if no ammunition
-   * @private
-   */
-  static async #checkAndConsumeAmmunition(actor, weapon) {
-    // Skip ammunition check for melee weapons
-    if (!weapon || !weapon.system.ranged) {
-      return true;
-    }
-
-    const loadedAmmoTypes = weapon.system.loadedAmmoTypes || {};
-    const hasAnyAmmunition = Object.values(loadedAmmoTypes).some(amount => amount > 0);
-    
-    if (!hasAnyAmmunition) {
-      ui.notifications.warn(`${weapon.name} não possui munição carregada!`);
-      return false;
-    }
-
-    // Consume 1 ammunition
-    const firstAmmoType = Object.keys(loadedAmmoTypes).find(type => loadedAmmoTypes[type] > 0);
-    if (firstAmmoType) {
-      const newAmount = loadedAmmoTypes[firstAmmoType] - 1;
-      await weapon.update({
-        [`system.loadedAmmoTypes.${firstAmmoType}`]: newAmount
-      });
-    }
-
-    return true;
-  }
-
-  /**
-   * Detect critical results for skill rolls
-   * @param {Roll} roll - The roll to analyze
-   * @param {Actor} actor - The actor
-   * @param {string} attributeKey - The attribute key (e.g., 'accuracy')
-   * @returns {Object} Flags object with critical information
-   * @private
-   */
-  static #detectCriticalResults(roll, actor = null, abilityKey = null) {
-    if (!roll || !roll.dice || roll.dice.length === 0) return {};
-
-    try {
-      // Evaluate the roll if not already evaluated
-      if (!roll._evaluated) {
-        roll.evaluate({ async: false });
-      }
-
-      const flags = {};
-      
-      // Check for critical failure (total ≤ 1 or natural 1)
-      if (roll.total <= 1) {
-        flags.criticalFailure = true;
-        return { cardigan: flags };
-      }
-
-      // Check for natural 1 on d20
-      // Only check ACTIVE dice (not discarded by advantage/disadvantage)
-      const d20Die = roll.dice.find(die => die.faces === 20);
-      if (d20Die && d20Die.results && d20Die.results.length > 0) {
-        const hasNaturalOne = d20Die.results.some(result => 
-          result?.active !== false && result?.result === 1
-        );
-        if (hasNaturalOne) {
-          flags.criticalFailure = true;
-          return { cardigan: flags };
-        }
-      }
-
-      // Check for critical hit - different logic for accuracy vs other rolls
-      // Only check ACTIVE dice (not discarded by advantage/disadvantage)
-      if (d20Die && d20Die.results && d20Die.results.length > 0) {
-        // For accuracy rolls, use actor's criticalHit threshold
-        if (abilityKey === 'accuracy' && actor && actor.system?.details?.criticalHit) {
-          const criticalThreshold = actor.system.details.criticalHit;
-          // Check if any active die result is 20 or higher for natural critical
-          const hasNaturalCritical = d20Die.results.some(result => 
-            result?.active !== false && result?.result === 20
-          );
-          if (roll.total >= criticalThreshold || hasNaturalCritical) {
-            flags.criticalHit = true;
-            return { cardigan: flags };
-          }
-        }
-        // For all other rolls, critical hit when total is 20 or higher OR natural 20
-        else {
-          const hasNaturalTwenty = d20Die.results.some(result => 
-            result?.active !== false && result?.result === 20
-          );
-          if (roll.total >= 20 || hasNaturalTwenty) {
-            flags.criticalHit = true;
-            return { cardigan: flags };
-          }
-        }
-      }
-
-      return {};
-
-    } catch (error) {
-      console.warn("Error detecting critical results:", error);
-      return {};
     }
   }
